# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "eWorldCup Manager API",
    "description": "API for managing round-robin tournament schedules",
    "version": "v1"
  },
  "paths": {
...
    "/Player": {
      "get": {
        "tags": [
          "Player"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Player"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Player/{playerIndex}/schedule": {
      "get": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "playerIndex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Player/{playerIndex}/{round}": {
      "get": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "playerIndex",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "round",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRoundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Player/{uid}": {
      "delete": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePlayerRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlayerDto": {
        "required": [
          "id",
          "name",
          "uid"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerRoundResponse": {
        "required": [
          "opponent",
          "opponentIndex",
          "player",
          "playerIndex",
          "round"
        ],
        "type": "object",
        "properties": {
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "playerIndex": {
            "type": "integer",
            "format": "int32"
          },
          "player": {
            "type": "string",
            "nullable": true
          },
          "opponentIndex": {
            "type": "integer",
            "format": "int32"
          },
          "opponent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerScheduleItemDto": {
        "required": [
          "opponent",
          "opponentIndex",
          "round"
        ],
        "type": "object",
        "properties": {
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "opponentIndex": {
            "type": "integer",
            "format": "int32"
          },
          "opponent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerScheduleResponse": {
        "required": [
          "n",
          "player",
          "playerIndex"
        ],
        "type": "object",
        "properties": {
          "n": {
            "type": "integer",
            "format": "int32"
          },
          "playerIndex": {
            "type": "integer",
            "format": "int32"
          },
          "player": {
            "type": "string",
            "nullable": true
          },
          "schedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerScheduleItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayersResponse": {
        "required": [
          "playerCount"
        ],
        "type": "object",
        "properties": {
          "playerCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxRounds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}